# apiGateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway для микросервисного приложения вишлистов",
        "title": "Wishlist API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8095",
    "basePath": "/api/v1",
    "paths": {
...
        "/wishlists": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новый вишлист для текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wishlists"
                ],
                "summary": "Создать вишлист",
                "parameters": [
                    {
                        "description": "Данные вишлиста",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateWishlistRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Вишлист создан",
                        "schema": {
                            "$ref": "#/definitions/handlers.WishlistDTO"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/wishlists/public": {
            "get": {
                "description": "Возвращает список публичных вишлистов с пагинацией",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wishlists"
                ],
                "summary": "Список публичных вишлистов",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Размер страницы",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список вишлистов",
                        "schema": {
                            "$ref": "#/definitions/handlers.WishlistListDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/wishlists/user/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает все вишлисты указанного пользователя (публичные или все для владельца)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wishlists"
                ],
                "summary": "Получить вишлисты пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список вишлистов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.WishlistListDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/wishlists/{id}": {
            "get": {
                "description": "Возвращает вишлист по его ID (публичные доступны всем, приватные только владельцу)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wishlists"
                ],
                "summary": "Получить вишлист",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID вишлиста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вишлист найден",
                        "schema": {
                            "$ref": "#/definitions/handlers.WishlistDTO"
                        }
                    },
                    "401": {
                        "description": "Не авторизован для приватного вишлиста",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "404": {
                        "description": "Вишлист не найден",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет существующий вишлист (только владелец)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wishlists"
                ],
                "summary": "Обновить вишлист",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID вишлиста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateWishlistRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вишлист обновлен",
                        "schema": {
                            "$ref": "#/definitions/handlers.WishlistDTO"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет вишлист и все его предметы (только владелец)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wishlists"
                ],
                "summary": "Удалить вишлист",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID вишлиста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешное удаление"
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateWishlistRequestDTO": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "description": "Description Описание вишлиста\nmaxLength: 1000\nexample: Подарки на мой день рождения",
                    "type": "string"
                },
                "is_public": {
                    "description": "IsPublic Флаг публичности вишлиста\nexample: true",
                    "type": "boolean"
                },
                "title": {
                    "description": "Title Название вишлиста\nrequired: true\nminLength: 1\nmaxLength: 80\nexample: На день рождения",
                    "type": "string"
                }
            }
        },
        "handlers.ErrorDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error содержит строку с расшифровкой того что пошло не так\nminLength: 1\nmaxLength: 100\nexample: item not found",
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdateWishlistRequestDTO": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "description": "Description Описание вишлиста\nmaxLength: 1000\nexample: Подарки на мой день рождения",
                    "type": "string"
                },
                "is_public": {
                    "description": "IsPublic Флаг публичности вишлиста\nexample: true",
                    "type": "boolean"
                },
                "title": {
                    "description": "Title Название вишлиста\nrequired: true\nminLength: 1\nmaxLength: 80\nexample: На день рождения",
                    "type": "string"
                }
            }
        },
...
        "handlers.WishlistDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt Дата и время создания (Unix timestamp)\nformat: date-time\nexample: 1700000000",
                    "type": "integer"
                },
                "description": {
                    "description": "Description Описание вишлиста\nmaxLength: 1000\nexample: Подарки на мой день рождения",
                    "type": "string"
                },
                "id": {
                    "description": "ID UUID вишлиста\nformat: uuid\nexample: 123e4567-e89b-12d3-a456-426614174000",
                    "type": "string"
                },
                "is_public": {
                    "description": "IsPublic Флаг публичности вишлиста\nexample: true",
                    "type": "boolean"
                },
                "owner_id": {
                    "description": "OwnerID UUID пользователя - владельца вишлиста\nformat: uuid\nexample: 123e4567-e89b-12d3-a456-426614174000",
                    "type": "string"
                },
                "title": {
                    "description": "Title Название вишлиста\nrequired: true\nminLength: 1\nmaxLength: 80\nexample: На день рождения",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt Дата и время последнего обновления (Unix timestamp)\nformat: date-time\nexample: 1700000000",
                    "type": "integer"
                }
            }
        },
        "handlers.WishlistListDTO": {
            "type": "object",
            "properties": {
                "current_page": {
                    "description": "CurrentPage Текущая страница\nexample: 1",
                    "type": "integer"
                },
                "total_count": {
                    "description": "TotalCount Общее количество вишлистов в списке Wishlists\nexample: 100",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "TotalPages Общее количество страниц\nexample: 10",
                    "type": "integer"
                },
                "wishlists": {
                    "description": "Wishlists Список вишлистов",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.WishlistDTO"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}